+++ .gradle/config.properties
@@ -1,2 +1,2 @@
-#Sat Jun 21 19:01:16 CST 2025
-java.home=C\:\\Program Files\\Android\\Android Studio\\jbr
+#Thu Jul 03 15:50:01 CST 2025
+java.home=E\:\\Android\\jbr
