+++ local.properties
@@ -1,10 +1,8 @@
-## This file is automatically generated by Android Studio.
-# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
-#
-# This file should *NOT* be checked into Version Control Systems,
+## This file must *NOT* be checked into Version Control Systems,
 # as it contains information specific to your local configuration.
 #
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-sdk.dir=C\:\\Users\\cl21d\\AppData\\Local\\Android\\Sdk
+#Thu Jul 03 15:50:07 CST 2025
+sdk.dir=C\:\\Users\\josep\\AppData\\Local\\Android\\Sdk
